Fix Java 17 and some OSGI startup pb
