+++ pom.xml
@@ -19,6 +19,7 @@
         <developerConnection>scm:git:git@github.com:tdraier/tmdb-provider.git</developerConnection>
     </scm>
     <properties>
+        <java.version>17</java.version>
         <require-capability>osgi.extender;filter:="(osgi.extender=org.jahia.bundles.blueprint.extender.config)"</require-capability>
         <jahia-depends>default,bootstrap3-components,bootstrap3-core,external-provider</jahia-depends>
         <jahia-module-signature>MCwCFCR4OLwEIo3wT7hHEhcxCLRp+DJqAhQ+Rxe43QQEVcO1O8lMQ+uxoHSWgw==</jahia-module-signature>
