+++ src/main/java/org/jahia/modules/provider/tmdb/item/mapper/MovieReferenceItemMapper.java
@@ -21,28 +21,37 @@
  *
  * ==========================================================================================
  */
-package org.jahia.modules.provider.tmdb.item;
+package org.jahia.modules.provider.tmdb.item.mapper;
 
-import org.jahia.api.Constants;
+import org.apache.commons.lang.StringUtils;
 import org.jahia.modules.external.ExternalData;
 import org.jahia.modules.provider.tmdb.helper.Naming;
 import org.jahia.modules.provider.tmdb.helper.PathBuilder;
+import org.jahia.modules.provider.tmdb.helper.PathHelper;
+import org.jahia.modules.provider.tmdb.item.ItemMapper;
+import org.jahia.modules.provider.tmdb.item.ItemMapperDescriptor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
-import java.util.*;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
 /**
- * Handler for persons node.
+ * Short description of the class
  *
  * @author Jerome Blanchard
  */
-@ItemMapperDescriptor(pathPattern = "^/persons$", idPattern = "^persons$", supportedNodeType = {Naming.NodeType.CONTENT_FOLDER},
-        hasLazyProperties = false)
-public class PersonsItemMapper extends ItemMapper {
+@ItemMapperDescriptor(pathPattern = "^/lists/[a-z0-9]+/\\d+$", idPattern = "^mref-[a-z0-9]+\\d+$", supportedNodeType =
+        {Naming.NodeType.CONTENT_REFERENCE}, hasLazyProperties = false)
+public class MovieReferenceItemMapper extends ItemMapper {
 
-    public static final String PATH_LABEL = "persons";
-    public static final String ID_PREFIX = "persons";
+    private static final Logger LOGGER = LoggerFactory.getLogger(MovieReferenceItemMapper.class);
+    public static final String PATH_LABEL = "lists";
+    public static final String ID_PREFIX = "mref-";
 
-    public PersonsItemMapper() {
+    public MovieReferenceItemMapper() {
     }
 
     @Override public List<String> listChildren(String path) {
@@ -50,14 +59,17 @@ public class PersonsItemMapper extends ItemMapper {
     }
 
     @Override public ExternalData getData(String identifier) {
+        String cleanId = StringUtils.substring(identifier, ID_PREFIX.length());
+        String listId = StringUtils.substringBefore(cleanId, "-");
+        int movieId = Integer.parseInt(StringUtils.substringAfter(cleanId, "-"));
         Map<String, String[]> properties = new HashMap<>();
-        properties.put(Constants.JCR_TITLE, new String[] {PATH_LABEL});
-        String path = new PathBuilder(PATH_LABEL).build();
-        return new ExternalData(identifier, path, Naming.NodeType.CONTENT_FOLDER, properties);
+        properties.put("j:node", new String[] { MovieItemMapper.ID_PREFIX + movieId });
+        String path = new PathBuilder(PATH_LABEL).append(listId).append(movieId).build();
+        return new ExternalData(identifier, path, Naming.NodeType.CONTENT_REFERENCE, properties);
     }
 
     @Override public String getIdFromPath(String path) {
-        return ID_PREFIX;
+        return ID_PREFIX.concat(PathHelper.getLeaf(path));
     }
 
     @Override public String getPathLabel() {
