+++ src/main/java/org/jahia/modules/provider/tmdb/item/mapper/MovieYearItemMapper.java
@@ -21,44 +21,58 @@
  *
  * ==========================================================================================
  */
-package org.jahia.modules.provider.tmdb.item;
+package org.jahia.modules.provider.tmdb.item.mapper;
 
+import org.apache.commons.lang.StringUtils;
 import org.jahia.api.Constants;
 import org.jahia.modules.external.ExternalData;
 import org.jahia.modules.provider.tmdb.helper.Naming;
 import org.jahia.modules.provider.tmdb.helper.PathBuilder;
+import org.jahia.modules.provider.tmdb.helper.PathHelper;
+import org.jahia.modules.provider.tmdb.item.ItemMapper;
+import org.jahia.modules.provider.tmdb.item.ItemMapperDescriptor;
 
 import java.util.*;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 
 /**
- * Handler for root node of the provider.
- * List predefined children based on root collections exposed
+ * Handler for movies years node.
+ * List month of the year limiting to current month if parent is current year
  *
  * @author Jerome Blanchard
  */
-@ItemMapperDescriptor(pathPattern = "^/$", idPattern = "^root$", supportedNodeType = {Naming.NodeType.CONTENT_FOLDER},
-        hasLazyProperties = false)
-public class RootItemMapper extends ItemMapper {
+@ItemMapperDescriptor(pathPattern = "^/movies/\\d{4}$", idPattern = "^myear-\\d{4}$", supportedNodeType =
+        {Naming.NodeType.CONTENT_FOLDER}, hasLazyProperties = false)
+public class MovieYearItemMapper extends ItemMapper {
+    public static final String ID_PREFIX = "myear-";
 
-    private static final List<String> CHILDREN = Arrays.asList(MoviesItemMapper.PATH_LABEL, PersonsItemMapper.PATH_LABEL);
-    public static final String ID_PREFIX = "root";
-
-    public RootItemMapper() {
+    public MovieYearItemMapper() {
     }
 
     @Override public List<String> listChildren(String path) {
-        return CHILDREN;
+        String node = PathHelper.getLeaf(path);
+        if (node != null) {
+            Calendar calendar = Calendar.getInstance();
+            int limit = (node.equals(Integer.toString(calendar.get(Calendar.YEAR)))) ? calendar.get(Calendar.MONTH) : 12;
+            return IntStream.rangeClosed(1, limit)
+                    .boxed().sorted(Collections.reverseOrder())
+                    .map(i -> node.concat("-").concat(StringUtils.leftPad(Integer.toString(i), 2, "0")))
+                    .collect(Collectors.toList());
+        }
+        return Collections.emptyList();
     }
 
     @Override public ExternalData getData(String identifier) {
+        final String year = StringUtils.substring(identifier, ID_PREFIX.length());
         Map<String, String[]> properties = new HashMap<>();
-        properties.put(Constants.JCR_TITLE, new String[] { "TMDB" });
-        String path = new PathBuilder().build();
+        properties.put(Constants.JCR_TITLE, new String[] { year });
+        String path = new PathBuilder(MoviesItemMapper.PATH_LABEL).append(year).build();
         return new ExternalData(identifier, path, Naming.NodeType.CONTENT_FOLDER, properties);
     }
 
     @Override public String getIdFromPath(String path) {
-        return ID_PREFIX;
+        return ID_PREFIX.concat(PathHelper.getLeaf(path));
     }
 
     @Override public String getPathLabel() {
