+++ src/main/resources/META-INF/services/org.jahia.modules.provider.tmdb.item.ItemMapper
@@ -1,10 +1,10 @@
-org.jahia.modules.provider.tmdb.item.ErrorItemMapper
-org.jahia.modules.provider.tmdb.item.RootItemMapper
-org.jahia.modules.provider.tmdb.item.PersonsItemMapper
-org.jahia.modules.provider.tmdb.item.PersonItemMapper
-org.jahia.modules.provider.tmdb.item.MoviesItemMapper
-org.jahia.modules.provider.tmdb.item.MovieYearItemMapper
-org.jahia.modules.provider.tmdb.item.MovieMonthItemMapper
-org.jahia.modules.provider.tmdb.item.MovieItemMapper
-org.jahia.modules.provider.tmdb.item.MovieCreditsItemMapper
-org.jahia.modules.provider.tmdb.item.MovieReferenceItemMapper
+org.jahia.modules.provider.tmdb.item.mapper.ErrorItemMapper
+org.jahia.modules.provider.tmdb.item.mapper.RootItemMapper
+org.jahia.modules.provider.tmdb.item.mapper.PersonsItemMapper
+org.jahia.modules.provider.tmdb.item.mapper.PersonItemMapper
+org.jahia.modules.provider.tmdb.item.mapper.MoviesItemMapper
+org.jahia.modules.provider.tmdb.item.mapper.MovieYearItemMapper
+org.jahia.modules.provider.tmdb.item.mapper.MovieMonthItemMapper
+org.jahia.modules.provider.tmdb.item.mapper.MovieItemMapper
+org.jahia.modules.provider.tmdb.item.mapper.MovieCreditsItemMapper
+org.jahia.modules.provider.tmdb.item.mapper.MovieReferenceItemMapper
